% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_logo_text.R
\name{add_text_logo}
\alias{add_text_logo}
\title{Add Stylized Cone Health to a ggplot figure}
\usage{
add_text_logo(fig, caption = NULL, captionsize = NULL)
}
\arguments{
\item{fig}{the ggplot item to apply the text logo to}

\item{caption}{optional, additional caption text to apply to the figure.
Note: for linebreaks in your caption, use \code{<br>} instead of
\code{\\n}.}

\item{captionsize}{optional text size in pts of the caption. The default of
\code{NULL} will use the default caption size of ggplot theme being used.}
}
\description{
Add the Text Logo to the ggplot2 figure.

\strong{Note:} This will override the use of \code{labs(caption = )}.
}
