% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_css.R
\name{apply_cone_css}
\alias{apply_cone_css}
\title{Applying  css Files}
\usage{
apply_cone_css(
  css_style = c("report", "shiny", "web", "guidebook", "presentation"),
  location = here::here(),
  new_name = NULL,
  combine = T,
  shiny_small = F
)
}
\arguments{
\item{css_style}{chacter vector identifying the type(s) of output that the
css is needed for. One or more of \code{shiny}, \code{web},
\code{report}, \code{presentation}, or \code{guidebook}
for a shiny web application, web page, rmarkdown report, presentation,
and/or guidebooks respectively.}

\item{location}{file path location of where to save the css file(s), defaults
to \code{here::here()}}

\item{new_name}{a character vector specifying the name(s) of the css file(s).
Should be of length 1 for \code{combine=T} (default) or the same length
of css_styles for \code{combine = F}.
The default of \code{NULL} will create a new file of the name
\code{styles.css} if combine = T. If combine = F each file requested will
be saved off separately and the default file names of \code{shiny.css},
\code{web.css}, \code{report.css}, etc. will be used}

\item{combine}{logical whether or not to combine the css files if multiple
are requested. Defaults to TRUE.}

\item{shiny_small}{logical indicating whether or not the shiny application
font should be smaller or not. Defaults to \code{FALSE}. Ignored for
\code{css_style!='shiny'}}
}
\description{
This function is designed to pull css files from the
\pkg{eastyle} package. Css files are available for shiny apps, websites,
rmarkdown reports, and more. These style sheets are intended to keep with
Cone Health brand standards.
The function returns a character vector containing the full file names
of all newly created css files on the users local machine.
}
