% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{brand_palettes}
\alias{brand_palettes}
\title{Return function to interpolate a brand color palette}
\format{
An object of class \code{list} of length 2.
}
\usage{
brand_palettes
}
\arguments{
\item{palette}{Character name of palette in pdw_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
This function creates a couple of palettes that could be used when plotting
}
\keyword{datasets}
