% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_logo_text.R
\name{cone_logo_text}
\alias{cone_logo_text}
\title{Generate Cone Health Stylized Text Logo}
\usage{
cone_logo_text(background = "#F9F9F9")
}
\arguments{
\item{background}{Background color of the logo text. Defaults to
\code{#F9F9F9}, the default background color of charts}
}
\description{
Creates a grid object with Cone Health Text Logo
}
