% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_main}
\alias{palette_main}
\title{Function to extract pdw colors as hex codes}
\format{
An object of class \code{character} of length 3.
}
\usage{
palette_main
}
\arguments{
\item{...}{Character names of pdw_colors}
}
\description{
This function creates the output vector for the colors into hex codes
}
\keyword{datasets}
