% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reactable.R
\name{reactable}
\alias{reactable}
\title{reactable Updated Defaults}
\usage{
reactable(data, striped = T, highlight = T, fontSize = "16px", ...)
}
\arguments{
\item{data}{the data}

\item{striped}{whether or not to stripe the columns, default \code{TRUE}}

\item{highlight}{whether or not to highlight entire row on hover, default \code{TRUE}}

\item{fontSize}{default \code{'14px'}}

\item{...}{with other options}
}
\description{
Reactable with striped rows, highlight on hover, and Cone Health
color and font.
}
