% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reactable.R
\name{reactableThemeCone}
\alias{reactableThemeCone}
\title{Reactable Theme with Cone Colors}
\usage{
reactableThemeCone(
  borderColor = "#dfe2e5",
  stripedColor = "#f6f8fa",
  highlightColor = "#f0f5f9",
  cellPadding = "8px 12px",
  headerBackgroundColor = "#00A2B2",
  headerColor = "#FFFFFF",
  fontFamily = "Montserrat, Arial, sans-serif",
  fontSize = "16px",
  ...
)
}
\arguments{
\item{borderColor}{border color}

\item{stripedColor}{striped columns color}

\item{highlightColor}{color of row on highlight}

\item{cellPadding}{table cell padding}

\item{headerBackgroundColor}{header background color}

\item{headerColor}{header font color}

\item{fontFamily}{table font}

\item{fontSize}{table font size}

\item{...}{other parameters passed to \code{reactableTheme}}
}
\description{
Default theme for reactables. Applied automatically when calling
\code{eastyle::reactable}
}
