% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tanger_colors}
\alias{tanger_colors}
\title{Cone Colors}
\format{
An object of class \code{character} of length 3.
}
\usage{
tanger_colors
}
\description{
The primary colour guide for Cone Health
}
\keyword{datasets}
