% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tanger_colors_pal1}
\alias{tanger_colors_pal1}
\title{Cone Colors 1}
\format{
An object of class \code{character} of length 3.
}
\usage{
tanger_colors_pal1
}
\description{
Secondary Themes
}
\keyword{datasets}
