% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teambrand-package.R
\docType{package}
\name{TeamBrand-package}
\alias{TeamBrand}
\alias{TeamBrand-package}
\title{Enterprise Analytics Styles}
\description{
An easy way to apply Tanger styles to Figures
and Documentation following the Tanger Identity Guidelines
}
\author{
\strong{Maintainer}: Philip Walker \email{philwalker2012@gmail.com}

}
\keyword{internal}
