% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_brand_defaults.R
\name{undo_brand_defaults}
\alias{undo_brand_defaults}
\title{Resets all the ggplot2 theme to its default settings}
\usage{
undo_brand_defaults()
}
\description{
Resets all the ggplot2 theme to its default settings
}
